ref: clean up and apply formatter
