+++ usvm-jvm-mocks/src/main/kotlin/machine/instructions/UTestMockConfigInfo.kt
@@ -3,5 +3,3 @@ package machine.instructions
 import org.usvm.test.api.UTestInst
 
 class UTestMockConfigInfo(val instructions: List<Pair<UTestInst, String>>)
-
-class UTestMockConfigInfo2(val instructions: List<UTestInst>)
