+++ usvm-jvm-mocks/src/main/kotlin/machine/render/ConfigInfoRenderer.kt
@@ -3,10 +3,8 @@ package machine.render
 import com.github.javaparser.ast.body.MethodDeclaration
 import com.github.javaparser.ast.expr.AnnotationExpr
 import com.github.javaparser.ast.expr.SimpleName
-import com.github.javaparser.printer.DefaultPrettyPrinter
 import machine.instructions.UTestMockConfigInfo
 import org.jacodb.api.jvm.JcClasspath
-import org.usvm.jvm.rendering.ReflectionUtilsInlineStrategy
 import org.usvm.jvm.rendering.baseRenderer.JcIdentifiersManager
 import org.usvm.jvm.rendering.baseRenderer.JcImportManager
 import org.usvm.jvm.rendering.testRenderer.JcTestVisitor
@@ -16,62 +14,7 @@ import org.usvm.jvm.rendering.unsafeRenderer.JcUnsafeUtilsRenderer
 import org.usvm.test.api.UTest
 import org.usvm.test.api.UTestExpression
 import org.usvm.test.api.UTestInst
-import org.usvm.test.api.UTestMethodCall
 import org.usvm.test.api.UTestMockInst
-
-fun renderConfigInfo(cp: JcClasspath, test: UTest, mockConfigInfo: UTestMockConfigInfo): String {
-    val importManager = JcImportManager()
-    val identifiersManager = JcIdentifiersManager()
-    val strategy = ReflectionUtilsInlineStrategy.NoInline()
-    val utilsRenderer = JcUnsafeUtilsRenderer(importManager, strategy)
-    val testClassRenderer = JcUnsafeTestClassRenderer(
-        "name",
-        importManager,
-        identifiersManager,
-        cp,
-        utilsRenderer
-    )
-    val testRenderer = ConfigInfoRenderer(mockConfigInfo, test, testClassRenderer, importManager, identifiersManager, cp, identifiersManager["test"], listOf(), utilsRenderer)
-    val res = testRenderer.render()
-    val printer = DefaultPrettyPrinter()
-    val text = printer.print(res)
-    val comments = testRenderer.renderConfigInfo()
-    return modifyText(text, reorder(comments))
-}
-
-fun reorder(lines: List<String>): List<String> {
-    val result = mutableListOf<String>()
-    val newlines = mutableListOf<String>()
-
-    for (line in lines) {
-        if (line == "\n") {
-            newlines.add(line)
-        } else {
-            result.add(line)
-            result.addAll(newlines)
-            newlines.clear()
-        }
-    }
-    result.addAll(newlines)
-    return result
-}
-
-
-fun modifyText(text: String, comments: List<String>): String {
-    val lines = text.split("\n") as MutableList<String>
-    lines.removeAt(lines.lastIndex)
-    lines.removeAt(lines.lastIndex)
-    lines.removeAt(0)
-    var finalText = ""
-    for (i in 0 until lines.size ) {
-        if (comments[i] != "\n") {
-            finalText = finalText + "//" + comments[i] + lines[i] + "\n"
-        } else {
-            finalText = finalText + lines[i] + "\n"
-        }
-    }
-    return finalText
-}
 
 class ConfigInfoRenderer(
     private val mockConfigInfo: UTestMockConfigInfo,
@@ -98,8 +41,9 @@ class ConfigInfoRenderer(
             return !preventVarDeclarationOf(expr) && isVisited(expr) || requireVarDeclarationOf(expr)
         }
         override fun visitExpr(expr: UTestExpression) {
-            if (shouldDeclareVarCheck(expr))
+            if (shouldDeclareVarCheck(expr)) {
                 shouldDeclareVar.add(expr)
+            }
 
             super.visitExpr(expr)
         }
