+++ usvm-jvm-mocks/src/test/kotlin/org/usvm/samples/ServiceTest.kt
@@ -4,13 +4,6 @@ import org.junit.jupiter.api.Test
 import org.usvm.test.util.checkers.ignoreNumberOfAnalysisResults
 
 class ServiceTest : MocksTestRunner() {
-//    override fun createMachine(
-//        cp: JcClasspath,
-//        options: UMachineOptions,
-//        interpreterObserver: JcInterpreterObserver?
-//    ): JcMachine {
-//        return JcMocksMachine(cp, options, interpreterObserver = interpreterObserver)
-//    }
 
     @Test
     fun testCalc() {
